+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/concurrent/TimerTaskPermissions.java
@@ -1,0 +1,10 @@
+package io.tech1.framework.domain.concurrent;
+
+import lombok.Data;
+
+// Lombok
+@Data(staticConstructor = "of")
+public class TimerTaskPermissions {
+    private final boolean start;
+    private final boolean stop;
+}