+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - Fix bug with @Mandatory + "get" in the middle of the property: target1 -> getTarget1() is expected getter, not tar1()
 - Migrate collection utility methods: get mutable list, set and map
 - Migrate tuples: TupleToggle + serialize/deserialize process improved
+- Add state machine -> TimerTask1 (created, operative, stopped)
 
 ### v1.6
 - Add "countryCode" -> GeoLocation